refactor: move annotation to types/xy_chart/annotations
