+++ src/components/annotation_tooltips.tsx
@@ -2,7 +2,7 @@ import { inject, observer } from 'mobx-react';
 import React from 'react';
 import { isLineAnnotation } from '../lib/series/specs';
 import { AnnotationId } from '../utils/ids';
-import { AnnotationDimensions, AnnotationLineProps } from '../state/annotation_utils';
+import { AnnotationDimensions, AnnotationLineProps } from '../types/xy_chart/annotations/annotation_utils';
 import { ChartStore } from '../state/chart_state';
 
 interface AnnotationTooltipProps {
