+++ src/components/react_canvas/reactive_chart.tsx
@@ -5,7 +5,11 @@ import { Layer, Rect, Stage } from 'react-konva';
 import { isLineAnnotation, isRectAnnotation, Position } from '../../lib/series/specs';
 import { LineAnnotationStyle, RectAnnotationStyle } from '../../utils/themes/theme';
 import { AnnotationId } from '../../utils/ids';
-import { AnnotationDimensions, AnnotationLineProps, AnnotationRectProps } from '../../state/annotation_utils';
+import {
+  AnnotationDimensions,
+  AnnotationLineProps,
+  AnnotationRectProps,
+} from '../../types/xy_chart/annotations/annotation_utils';
 import { ChartStore, Point } from '../../state/chart_state';
 import { BrushExtent } from '../../state/utils';
 import { AreaGeometries } from './area_geometries';
